requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# COCO API
pycocotools>=2.0.7

# Data / numerics
numpy>=1.24.0
Pillow>=10.0.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0
